Refactor version update
